// apps/frontend-v2/src/lib/api/model/inboxPatientDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Lean Commerce API
 * The AI Sales Agent CRM API Documentation
 * OpenAPI spec version: 1.0
 */

export interface InboxPatientDto {
  id: string;
  firstName: string;
  lastName: string;
  phoneNumber: string;
  /** @nullable */
  email: string | null;
  primaryLanguage: string;
  country: string;
  timezone: string;
  status: string;
  priority: string;
  /** @nullable */
  assignedAgentId: string | null;
  /** @nullable */
  assigneeName: string | null;
  /** @nullable */
  pipelineStageId: string | null;
  /** @nullable */
  stageName: string | null;
  /** @nullable */
  summary: string | null;
  optedOut: boolean;
}